Update CodeCoverage.cmake
The unpacking of the _testrunner argument allows to use test runners that cannot be specified as a single command.
MPI applications have to be executed indirectly via mpirun, for example:

setup_target_for_coverage(coverage "mpirun -n 4 ${TEST_TARGET}" test_coverage)

The mode UNIX_COMMAND is not platform-dependent. It is specified so arguments are separated by spaces instead of semicolons.